+++ test/test_install_report/install_report_test.sh
@@ -15,15 +15,16 @@ oneTimeSetUp() {
 }
 
 testBundling() {
-    ./node_modules/.bin/parcel build --no-scope-hoist --no-cache > "$DIR"/artifacts/testBundling.txt 2>&1
+    yarn run build > "$DIR"/artifacts/testBundling.txt 2>&1
 
     output=$(cat "$DIR"/artifacts/testBundling.txt)
-    line1=$(sed '1q;d' "$DIR"/artifacts/testBundling.txt)
 
-    assertEquals "Building..." "$line1"
-    assertContains "$output" "Bundling..."
-    assertContains "$output" "Packaging & Optimizing..."
-    assertContains "$output" "✨ Built in"
+    assertContains "$output" "main.js"
+    assertContains "$output" "main.css"
+    assertContains "$output" "service-worker.js"
+    assertContains "$output" "Done in"
+    assertNotContains "$output" "fail"
+    assertNotContains "$output" "error"
 }
 
 testInstall() {
@@ -52,7 +53,7 @@ testWeeReportRunAlternative() {
     assertContains "$output" "Using configuration file /home/weewx/weewx.conf"
     assertContains "$output" "Generating for all time"
     assertContains "$output" "INFO weewx.cheetahgenerator: Generated 43 files for report WdcReport in"
-    assertContains "$output" "INFO weewx.reportengine: Copied 9 files to /home/weewx/public_html"
+    assertContains "$output" "INFO weewx.reportengine: Copied 10 files to /home/weewx/public_html"
 
     assertNotContains "$output" "failed with exception"
     assertNotContains "$output" "Ignoring template"
@@ -68,7 +69,7 @@ testWeeReportRunClassic() {
     assertContains "$output" "Using configuration file /home/weewx/weewx.conf"
     assertContains "$output" "Generating for all time"
     assertContains "$output" "INFO weewx.cheetahgenerator: Generated 43 files for report WdcReport in"
-    assertContains "$output" "INFO weewx.reportengine: Copied 9 files to /home/weewx/public_html"
+    assertContains "$output" "INFO weewx.reportengine: Copied 10 files to /home/weewx/public_html"
 
     assertNotContains "$output" "failed with exception"
     assertNotContains "$output" "Ignoring template"
@@ -83,7 +84,7 @@ testWeeReportRunWithoutWeewxForecast() {
     assertContains "$output" "Using configuration file /home/weewx/weewx.conf"
     assertContains "$output" "Generating for all time"
     assertContains "$output" "INFO weewx.cheetahgenerator: Generated 43 files for report WdcReport in"
-    assertContains "$output" "INFO weewx.reportengine: Copied 9 files to /home/weewx/public_html"
+    assertContains "$output" "INFO weewx.reportengine: Copied 10 files to /home/weewx/public_html"
 
     assertNotContains "$output" "failed with exception"
     assertNotContains "$output" "Ignoring template"